Update Form1.cs
Hi! I´ve changed the name of the list because is the same that a variable. I will need to change in the other files! I´ve change too your strings for StringBuilder. string in C3 are inmutable, which means that if you add a char, you are making a new copy of the string. But with StringBuilder it doesn´t happend.
Best regards! Nice job